v1.5.1: update cxf and seclib versions; allow to set a validator

+++ src/test/java/eu/unicore/security/wsutil/TestClient.java
@@ -26,7 +26,7 @@ public class TestClient extends AbstractTestBase{
 		assertEquals("pass",ap.getPassword());
 
 		assertEquals("http://foo",hc.getClient().getProxyServer());
-		assertEquals(123,hc.getClient().getProxyServerPort());
+		assertEquals(123,(int)hc.getClient().getProxyServerPort());
 		
 	}
